handle named bundles
